+++ src/components/usa-button/usa-button.stories.js
@@ -1,3 +1,6 @@
+import { SystemColorList, ThemeColorList, ColorListAll } from "../../tokens/color/color~vars";
+import modifierList from './button';
+import Docs from "./docs/index.mdx";
 import Component from "./usa-button.twig";
 import {
   Data,
@@ -11,8 +14,33 @@ import {
   UnstyledData,
 } from "./content";
 
+console.log(ColorListAll);
+
 export default {
   title: "Components/Button",
+  argTypes: {
+    text: {
+      control: { type: 'text' },
+      defaultValue: 'Default',
+      name: 'Button text',
+    },
+    modifier: {
+      control: { type: 'select' },
+      name: 'Class modifier',
+      options: modifierList,
+    },
+    is_demo: {
+      control: { type: 'boolean' },
+      defaultValue: true,
+      name: 'Display all interactive states',
+    },
+  },
+  parameters: {
+    layout: 'fullscreen',
+    docs: {
+      page: Docs,
+    },
+  },
 };
 
 const Template = (args) => Component(args);
@@ -43,3 +71,115 @@ Secondary.args = SecondaryData;
 
 export const Unstyled = Template.bind({});
 Unstyled.args = UnstyledData;
+
+export const StylePlayground = Template.bind({});
+StylePlayground.argTypes = {
+  playground_mode: {
+    control: { type: 'boolean' },
+    defaultValue: 'true',
+    table: { disable: true }
+  },
+  modifier: { 
+    table: { disable: true },
+  },
+  display_element_states: { 
+    table: { disable: true },
+  },
+  /* $system-properties > $system-spacing */
+  border_radius: {
+    control: { type: 'select' },
+    defaultValue: 'md',
+    name: '$theme-button-border-radius',
+    options: {
+      "none": 0,
+      "sm": "2px",
+      "md": "4px",
+      "lg": "8px"
+    }
+  },
+  stroke_width: {
+    control: { type: 'text' },
+    defaultValue: '2px',
+    name: '$theme-button-stroke-width',
+  },
+  font_family: {
+    control: { type: 'text' },
+    defaultValue: '0',
+    name: '$theme-button-font-family',
+    options: {
+      
+    }
+  },
+  background_color: {
+    control: { type: 'select' }, 
+    defaultValue: 'blue-60v',
+    name: 'Default/focus state: background color',
+    options: SystemColorList,
+  },
+  text_color: {
+    control: { type: 'select' }, 
+    defaultValue: 'white',
+    name: 'Default/focus state: text color',
+    options: SystemColorList,
+  },
+  background_color_hover: {
+    control: { type: 'select' },
+    defaultValue: 'blue-warm-70',
+    mapping: SystemColorList,
+    name: 'Hover state: background color',
+    options: SystemColorList,
+  },
+  text_color_hover: {
+    control: { type: 'select' },
+    defaultValue: 'white',
+    mapping: SystemColorList,
+    name: 'Hover state: text color',
+    options: SystemColorList,
+  },
+  background_color_active: {
+    control: { type: 'select' },
+    defaultValue: 'blue-warm-70',
+    mapping: SystemColorList,
+    name: 'Active state: background color',
+    options: SystemColorList,
+  },
+  text_color_active: {
+    control: { type: 'select' },
+    defaultValue: 'white',
+    mapping: SystemColorList,
+    name: 'Active state: text color',
+    options: SystemColorList,
+  },
+  background_color_disabled: {
+    control: { type: 'select' },
+    defaultValue: 'blue-warm-70',
+    mapping: SystemColorList,
+    name: 'Disabled state: background color',
+    options: SystemColorList,
+  },
+  text_color_disabled: {
+    control: { type: 'select' },
+    defaultValue: 'white',
+    mapping: SystemColorList,
+    name: 'Disabled state: text color',
+    options: SystemColorList,
+  },
+  text_color_unstyled: {
+    control: { type: 'select' },
+    defaultValue: 'blue-warm-70',
+    mapping: SystemColorList,
+    name: 'Unstyled: text color',
+    options: SystemColorList,
+  },
+  font_size: {
+    control: { type: 'text' },
+    defaultValue: '1.06rem',
+    name: 'Font size',
+  },
+  stroke_color: {
+    control: { type: 'select' },
+    mapping: SystemColorList,
+    name: 'Stroke Color (stroke available on outline variant only)',
+    options: SystemColorList,
+  },
+}
